External Thrusters implementation
While it should have been an 'exotic system' rather than an upgrade from the start,
the upgrade thing works great. It's my new favorite thing. Most of the tiny bugs and kinks
have been ironed out.

Which means, it's time to make it a real exotic system now, and maybe fix the icon too.
The tooltip already works great, but perhaps the "skill" part of it could use more loving.

TO BE AMMENDED

+ faction config touchups (upgrade proliferation throughout a few files)