Ported over changes/upgrades from the other ExoticaTech 2.0.0rc4 mod
Since lacking the upgrade data in JSONs for the new things I already had (but my mod did not)
was the real root-cause of all the null-pointer shenanigans, doublebangs everywhere are still a bad thing.

Now that everything is ported over, and the mod works fine, the thing(s) that need to be looked into further
are concerning the ExternalThrusters activation by the AI. I'm not too sure the condition is good.

Still, it's good to go, and the 4 new upgrades show up fine. Whether they work fine is to be seen.

Also, bumped the version to 2.0.1