The AI uses the ability
While it seems there's a bug in the calculations too, at least the ability is being used.
Based on range alone.

targetDir: 83.03211, differenceInDegrees: 81.09238, distanceToTargetSquared: 6256002.0, activationDistanceSquared: 4000000, farEnoughToActivate: true

I need to figure out why the differenceInDegrees is so huge when the ship is aiming directly at the enemy.
I have an idea though. Could be a dumb semantical mistake. I need to verify.

EDIT:
after a slight ammend and a bugfix, i was right, my 'differenceInDegrees' was semantically wrong.
Now that it's fixed, and useful code extracted to extension functions and the whole method refactored,
it's working as it should have been from the start. Hooray.