The shield is functional, the shield-hosting drone is no longer being shot down
However, the constant 'ticking' sound and the floating numbers lead me to think that
there is a consistent and constant collision between the shieldhost and the owner ship.

Need to look into that, and the fact that the owner-ship seems to rotate quite a bit
or faster than it normally does - however i'm not entirely sure about this part.

The guardianshield drone was changed and had it's bounds reduced to a small 2x2 square.
It is also using a new transparent square sprite, so it's no longer visible on the ownership.
This could be looked into for further improvement.

In any case, the collision channel needs to be looked into before further refinements are made.