Fix the shield-host / owner-ship collision
Changing the shield-host (drone) collision channel to FIGHTER seems to have solved the problem.
Hooray.

Next thing up - the actual dynamic of how the shield opens up.

The Mammut's shield starts as a full circle that opens outwards by increasing radius.
The GuardianShield opens like a normal shield, from the center of the arc and then expands to the sides.

I think that's the last hurdle before the small polish starts, like the description string, the TB cost and that sorta thing.