Shield opening outwards - done
Rather than the old "shield opening by spreading it's arc to 360-degrees" now it
starts at 360 degrees and opens outwards by increasing it's radius.

I hope/expect that both of the newly added things are functional as well -
the force moving stuff out of the shield and the outward-opening of the shield
work fine.

Time to move to low-hanging fruit such as the description, the TB cost and
and similar polish.