Changing the drone to FXDrone
Well, this is both better and worse than the previous droid.

Now we're stuck with a giant triangle for the FX drone, but there's no
'outline' that outlines frigates centered on the ship with it's name.

So, win some, lose some. There is one more thing to try that might work though
- modular ship, and not rendering the part with the shield. That way it might
avoid the whole triangle.

Since modular stations are just one thing, if the module owns the shield then
it won't grow big on the tacmap, which is kinda a success.