Hotfix for GuardianShield on ships without a shield
Fix the case when the GuardianShield is installed on a ship without a shield,
and the "save/restore original shield" mechanic which didn't take into account
the case when the ship won't have a shield, and kotlin's non-null member would throw
and cause a crash.

Fixed by saving the original shield only if it's non-null, and restoring original shield
only if it was saved. Also tested whether the 'original shield' is actually persisted
through saved games or different runs of the game, and yes, everything checks out and works fine.

EXPECTED BEHAVIOR:
- Installing GuardianShield on a ship with a shield removes the ship's original shield
- Installing GuardianShield on a ship without a shield effectivelly does nothing regarding it's null shield
- Removing GuardianShield on a ship with a shield restored the original shield
- Removing GuardianShield on a ship without a shield effectivelly does nothing regarding it's null shield

BEFORE FIX:
- Installing GuardianShield on a ship with a shield removed the original shield
- Installing GuardianShield on a ship without a shield crashes the game
- Removing GuardianShield on a ship with a shield restored the original shield
- Removing GuardianShield on a ship without a shield was untestable due to crash

AFTER FIX:
- Installing GuardianShield on a ship with a shield removed the original shield
- Installing GuardianShield on a ship without a shield does nothing regarding it's null shield
- Removing GuardianShield on a ship with a shield restored the original shield
- Removing GuardianShield on a ship without a shield does nothing regarding it's null shield