SynchronizedRingBuffer is covered with tests and works alright even in multithread mode
Well, so far the idea is to have all the same tests as the base RingBuffer
plus a few multithreaded ones.

However, after the first multithreaded test was done, the necessity for more than one
becomes questionable, because one thread is producing, one is consuming, and while
items aren't quite made-then-produced at the same time, they are always consumed
in 1-10 order even if multiple items were produced between two get()s.

I wanted to put the search into the mix, but there's no way to test that while they're
both produced and consumed at the same time. So a test that just produces and searches
exists and works fine :)