Well now, this is embarassing... (it works)
Apart from the whole location.set(oldLocation) which might have been broken,
the mod might have worked all along. The system never activated because the
autopilot was driving the ship, not me...

Either way, after confirming it works, I am dissatisfied with how it actually
works. I want the system to be phased-in and phased-out of, and not for it to be
an immediate, visual-less teleportation.

So, the activation system needs reworking.