New "Rewind System" and a bunch of other goodies (#12)
* New 'datastructures' package

It's a new place to dump necessary data structures instead of just keeping them local to relevant features as inner classes.

For now there's the new (overly parametrized) generic RingBuffer and a similar
multi-producer multi-consumer thread-safe SynchronizedRingBuffer.

* Add support for tests :D

Well, the datastructures need tests so we know they work as they should work.
And it's a good thing i added them too, because I'm seeing alot of potential
problems between kotlin/java types, namely Integer, kotlin Int and casting to Object ...

So some special care needs to be put in into the createArray() most likely.

* RingBuffer is working correctly and is covered with tests

Perhaps the coverage isn't 100% but i'm glad with how it works and
how all functionalities are covered with tests.

Javadoc is now nice too.

* SynchronizedRingBuffer is covered with tests and works alright even in multithread mode

Well, so far the idea is to have all the same tests as the base RingBuffer
plus a few multithreaded ones.

However, after the first multithreaded test was done, the necessity for more than one
becomes questionable, because one thread is producing, one is consuming, and while
items aren't quite made-then-produced at the same time, they are always consumed
in 1-10 order even if multiple items were produced between two get()s.

I wanted to put the search into the mix, but there's no way to test that while they're
both produced and consumed at the same time. So a test that just produces and searches
exists and works fine :)

* Rudimentary Optional implementation

* Initial RewindSystem implementation

I think the most of the "meat" is on the bone. Some touchups probably needed but - should be testable.
Just need the activation/deactivation code in and it's good to be put into the game.

* Follow-up RewindSystem commit - it's ready for testing

the onActivate / onFinish were implemented as well, time to give it a spin.
Maybe it should have been in one place, in onStateSwitched() but that can be
changed after code review

* Added icons, the data/.../csv stuff and deployed for testing

Only the data/strings/string.json needs touching up, but all the other entries should be covered

Hope it works fine :)

* Replace plain reflection with MethodHandle reflection calls

tl;dr is - I had to use java.lang.reflection to bypass the "reified generic parameter" problem
so I used java.lang.reflect.Array.newInstance() to instantiate arrays of a given class and size,
and then casted them to concrete Array<T>.

However, starsector's security manager butted in and said "no reflection" so that had to be fixed.
Which is when the next problem happened - when I tried doing what I simply managed to do with JavaReflectionUtils
in Kotlin, the code would have ended up there. I didn't know how to reference the "int.class" primitive class that
is usable in Java but i dont know how to use it in Kotlin and my method signatures never matched.

The Java MethodHandles even pass all tests too.

Also, new JAR is put here too.

* Fix crash from last night with a missing comma in exotics.json

* Made the icon the same size as others; also gave it a facelift

* The dynamic string finally works great; the rewind system not so much

But hey, at least it doesn't crash anymore. Just doesn't work alright at all.

* Plenty of small bugfixes but it still doesn't work

Though i think i have a good idea why now.

* Make copies of saved values instead of just referencing originals

One of the problems was that it was just using the same reference
so that when it tried teleporting you somewhere from past, it
essentially "teleported" you to where you're currently standing.

* I broke it, it can't even activate now :(

* Well now, this is embarassing... (it works)

Apart from the whole location.set(oldLocation) which might have been broken,
the mod might have worked all along. The system never activated because the
autopilot was driving the ship, not me...

Either way, after confirming it works, I am dissatisfied with how it actually
works. I want the system to be phased-in and phased-out of, and not for it to be
an immediate, visual-less teleportation.

So, the activation system needs reworking.

* System activation rework - phase-in mechanic (WIP)

To avoid the system being instantaneous, i'm reworking it to be ~2sec long activation
before instant teleportation happens. This made the code somewhat messy as one thing
is now happening in multiple methods. The idea is sound, lets see what it looks like
during testing.

* Works great, ready for CR and merging-in

* Add RewindSystem to modSettings.json so it can spawn randomly

* Added playing sounds

* Get rid of backup; new JAR

deleted backup folder, added new jar

* Fix sounds; new JAR

* Address CR comments and a fresh JAR