Slight MissileSpamSystem rework
Well, it just occured to me that reloadDelay might not be the thing I want to look at
at all. So, I started messing with the refireDelay, and repurposed the old reloadDelay
into free-reloading.

Also, the shouldAffectWeapon() was tweaked, because we don't really care whether the
weapon uses ammo - but only if it's a missile. Let it deal with it's own ammo.

Ok, the new system idea is:
- only affects missile weapons, regardless of their ammo
- when system is activated it makes their refire rates much faster:
either 33% original or 1, whichever is lower
- those (missile) weapons that use ammo get 2 free reloads
- when the system deactivates, we get rid of the reloadMap