Completely redesigned/reworked the system and entirely simplified it
Instead of the previous "make missiles shoot out double/triple projectiles while the system is active, and give free reloads"
idea, the system is greatly simplified now.

* it doubles the missile rate-of-fire when installed
* when it's activated, it slaps on an additional x60 modifier on top of that

So, all it does now is just apply/deapply the active buff now.

3003157 [Thread-3] INFO  exoticatechnologies.modifications.exotics.impl.MissileSpamSystem  - [MissileSpamSystem] onActivate()	applied buff MissileSpamSystemRoFIncrease: buffed missileRoFMult to 150.0 from 2.5
3031655 [Thread-3] INFO  exoticatechnologies.modifications.exotics.impl.MissileSpamSystem  - [MissileSpamSystem] onFinished()	unapplied buff MissileSpamSystemRoFIncrease: restoring missileRoFMult to 2.5 from 150.0
3315619 [Thread-3] INFO  exoticatechnologies.modifications.exotics.impl.MissileSpamSystem  - [MissileSpamSystem] onActivate()	applied buff MissileSpamSystemRoFIncrease: buffed missileRoFMult to 150.0 from 2.5
3348651 [Thread-3] INFO  exoticatechnologies.modifications.exotics.impl.MissileSpamSystem  - [MissileSpamSystem] onFinished()	unapplied buff MissileSpamSystemRoFIncrease: restoring missileRoFMult to 2.5 from 150.0

It no longer needs to monitor all weapons and do all that other thing i thought of initially doing, for which i made the ringbuffer and all of that...
oh well, the ringbuffer willll be used for something eventually. the test harness will be useful as well.

but this system actually works fine as it is now. so it just needs cleanup.