Missile spam system (#13)
* Missile Spam system - initial commit

* Added to special_items.csv

* Added to exotics.json

* Added to settings and modSettings

* Add cargo/exotics icons

* Fill in the modifyTooltip() method and the latest JAR

* Add afterimage, debug logs and start messing with refireDelay

* Actually add to strings.json because i didn't save the file before...

* Slight MissileSpamSystem rework

Well, it just occured to me that reloadDelay might not be the thing I want to look at
at all. So, I started messing with the refireDelay, and repurposed the old reloadDelay
into free-reloading.

Also, the shouldAffectWeapon() was tweaked, because we don't really care whether the
weapon uses ammo - but only if it's a missile. Let it deal with it's own ammo.

Ok, the new system idea is:
- only affects missile weapons, regardless of their ammo
- when system is activated it makes their refire rates much faster:
either 33% original or 1, whichever is lower
- those (missile) weapons that use ammo get 2 free reloads
- when the system deactivates, we get rid of the reloadMap

* Update icon, get rid of background black color

* Update strings, update tooltip, new JAR

* Stop using copypasted logs and use the method that does this

* Stop using Weapon ID as teh refire map's key

Well... when the ship packs multiples of the same weapon - guess what happens?
that's right. only one of them ends up in the map, and teh rest remain unchanged.

So use teh WeaponAPI reference as the key instead of the weapon.id since hopefully
those won't end up overwriting each other.

* Completely redesigned/reworked the system and entirely simplified it

Instead of the previous "make missiles shoot out double/triple projectiles while the system is active, and give free reloads"
idea, the system is greatly simplified now.

* it doubles the missile rate-of-fire when installed
* when it's activated, it slaps on an additional x60 modifier on top of that

So, all it does now is just apply/deapply the active buff now.

3003157 [Thread-3] INFO  exoticatechnologies.modifications.exotics.impl.MissileSpamSystem  - [MissileSpamSystem] onActivate()	applied buff MissileSpamSystemRoFIncrease: buffed missileRoFMult to 150.0 from 2.5
3031655 [Thread-3] INFO  exoticatechnologies.modifications.exotics.impl.MissileSpamSystem  - [MissileSpamSystem] onFinished()	unapplied buff MissileSpamSystemRoFIncrease: restoring missileRoFMult to 2.5 from 150.0
3315619 [Thread-3] INFO  exoticatechnologies.modifications.exotics.impl.MissileSpamSystem  - [MissileSpamSystem] onActivate()	applied buff MissileSpamSystemRoFIncrease: buffed missileRoFMult to 150.0 from 2.5
3348651 [Thread-3] INFO  exoticatechnologies.modifications.exotics.impl.MissileSpamSystem  - [MissileSpamSystem] onFinished()	unapplied buff MissileSpamSystemRoFIncrease: restoring missileRoFMult to 2.5 from 150.0

It no longer needs to monitor all weapons and do all that other thing i thought of initially doing, for which i made the ringbuffer and all of that...
oh well, the ringbuffer willll be used for something eventually. the test harness will be useful as well.

but this system actually works fine as it is now. so it just needs cleanup.

* Updated the description string and some cleanup

* Turn off debug flag and more cleanup

* Even more cleanup

* New JAR

* Adding a deploy.sh script for deployment

I missed copying over the data folder and the new string for the last time.
I'm automating this deployment right now.

* Update .gitignore to ignore DEPLOYMENT folder

* Add highlighting to system description replacements

* Better description wording

EDIT: added hyphens for better emphasis

* Latest JAR

* New system title name color and afterimage color

Lets use the radioactive green for the system's name (description title)

but somehow, even after comparing screenshots, i dislike the radioactive green
used on the afterimage, and prefer how the Color.GREEN.brighter() looks.

so lets use that. my new color for the title, and the old color for afterimage.

EDIT: done. using RADIOACTIVE_GREEN for title, Color.GREEN.brighter() for afterimage