Unhardcode the exotic base price of 250000 in MarketMenu and let them define their prices
Well, the initial 'purchase' price for exotics was 250000. But it was hardcoded.

I want exotics to be able to define their own prices, in their "ExoticSystem() : Exotic"
class definition, instead of in the data CSV files.

So, having unhardcoded the price from 250k in MarketMenu, the ExoticSpecialItemPlugin was made
to read the price from it's exotic if possible (non-null) or fallback to 250k in case it is null.

This should technically allow me to declare prices in the Exotic system classes themselves and not use
workarounds to deduce some amount of money otherwise to make up for it.