Figure out a way to dig out the price set in the code in the Exotica Branch shop
So, in the purchase menu, for some reason the 'exotic' representing the exotic system
is always null, but it also seems that, whenever the 'exotic' is null, the exoticData is
non-null, and furthermore, exoticData.exotic is non-null and represents the exotic system
we're interacting-with with the mouse. So, just use that in the else part of figuring out
the cost.

In the end, if both exotic and exoticData are null, fallback to 250k. I can live with that.