Minor misc updates (#14)
Minor Misc Updates across the board

This PR is made to alleviate some grievances and usage warts that have been ticking me off, such as "not being able to set the exotic's selling price on a per-exotic-system basis from the code"

Additionally, it includes some other quality-enhancers such as getting rid of doublebanging in alot of places, which, while debatable, is still better than just doublebanging everywhere.

Features include:
- unhardcoding exotic price from 250000
- exotics can now set their own base price from the code by overriding the new `getBasePrice()` method
- upgrades don't have this method since their price calculation was always done differently (based on the level of the upgrade and number of stack items)
- no more doublebangs everywhere, we now properly use `?.let { ... }` eyesore semantics everywhere instead of that
- RewindSystem set it's price to 2500000
- Incorporated LunaSettings into the mod, made a setting for maximum number of exotic systems on the ship
- got rid of the "remove X cargo.items.credit" custom price workarounds from `canAfford()`/`removeItemsFromFleet()` which made custom prices a complete sideeffect of removing the custom (or chip) item from the inventory
- make all places use ETModPlugin.MOD_ID instead of hardcoding "exoticatechnologies" as mod ID everywhere
- ditto for the 250k fallback for exotic base prices
- Consolidate FullMetalSalvo to properly use `getPositiveMult` and `getNegativeMult`
- Consolidate MissileSpamSystem to properly use `getPositiveMult` and `getNegativeMult`
- Consolidate RewindSystem to properly use just `getNegativeMult` for cooldown since `getPositiveMult` has nothing to improve here
- update strings.json to include the "cooldown_string" and get rid of most of the "by lazy" string replacements since they're no longer necessary. they were before, but not anymore.
- HackedMissileForge costs 150k
- MissileSpamSystem costs 500k
- RewindSystem costs 2.5mil


trailing changelog from commits:

* Unhardcode the exotic base price of 250000 in MarketMenu and let them define their prices

Well, the initial 'purchase' price for exotics was 250000. But it was hardcoded.

I want exotics to be able to define their own prices, in their "ExoticSystem() : Exotic"
class definition, instead of in the data CSV files.

So, having unhardcoded the price from 250k in MarketMenu, the ExoticSpecialItemPlugin was made
to read the price from it's exotic if possible (non-null) or fallback to 250k in case it is null.

This should technically allow me to declare prices in the Exotic system classes themselves and not use
workarounds to deduce some amount of money otherwise to make up for it.

* Similar as last commit, but for upgrades and some improvements for exotics

* Remove more doublebangs from ExoticSpecialItemPlugin

* Change MissileSpamSystem's price to 500k from it's own code

And it doesn't use the old removeItemsFromFleet workaround (side-effect)
anymore.

* Change RewindSystem's price to 2.5mil from it's own code

* Slight update to Exotic's javadocs

* Figure out a way to dig out the price set in the code in the Exotica Branch shop

So, in the purchase menu, for some reason the 'exotic' representing the exotic system
is always null, but it also seems that, whenever the 'exotic' is null, the exoticData is
non-null, and furthermore, exoticData.exotic is non-null and represents the exotic system
we're interacting-with with the mouse. So, just use that in the else part of figuring out
the cost.

In the end, if both exotic and exoticData are null, fallback to 250k. I can live with that.

* Latest JAR with the new prices calculation in-place and working

* Make all references to 250000 use the same value

Since 250000 is the default base price of Exotics, it makes sense to put it
in the Exotic class too, and have all other places reference that value.

* Make all places use ETModPlugin.MOD_ID instead of hardcoding 'exoticatechnologies' as mod ID

* Add foundation for LunaSettings

* Consolidate remaining exoticas to use getBasePrice

Prefer using the getBasePrice() approach instead of deducting money in canAfford / removeItemsFromFleet
methods, since those might not even be visible to the player ahead of time to make an informed decision.

EDIT1: consolidate ingame naming to "Exoticas"
EDIT2: latest JAR

* Consolidate FullMetalSalvo to use getPositiveMult and getNegativeMult correctly

* Consolidate MissileSpamSystem to use getPositiveMult and getNegativeMult correctly

* Consolidate RewindSystem to use getPositiveMult and getNegativeMult correctly

Well, just getNegativeMult for the cooldown actually, there's nothing the positiveMult can
change here without messing things up.

* Consolidate MissileSpamSystem to use getPositiveMult and getNegativeMult correctly

And get rid of these "by lazy" string replacements and do some cleaning up while at it too

* Latest JAR with all the changes

* Fix strings.json, I forgot to prefix enclosed replacements things with $

* Address CR comments

* Address CR comments: add/improve javadocs