Get rid of exoticStuffHolder member, it's uses and relevant methods
Unfortunatelly this produced a small usage wart of reloadWeapon() having to be
extended with more arguments, but on the other hand - why not.

The class will survive for now, because it just *might* become useful for edge-cases
where we need to do something with those three things without having access to them.

But for now, it's nuked out of the system that brought it to existence :D