Add playing of sounds to AmmoCraftingSystem, some necessary refactoring and new JAR
Well, considering that RewindSystem had a playSound() method of it's own, it'd be kinda dumb
to just copypaste it and not reuse it. So it got moved to a KtxUtils file in the utils package.

So, RewindSystem was affected by that move, and had to be adjusted to use the new version of the method.
And AmmoCraftingSystem just added usages of the playSound() method from the KtxUtils as well.

Also added the 3 new sounds, and updated the sounds.json

EDIT: now i've added the sounds.json modifications but also - the new JAR as well