Fix a bug before it actually happens and I fail to catch it :)
Namely, as I was writing out the PR description, and I came to the part that says
"by default, period lasts 10 seconds (affected by positive/negative multipliers)"
i remembered that I hardcoded the IntervalUtil to last 9.95,10.05 seconds.

In other words, if somehow the period got to last 5 seconds, this thing would
only tick every 10 seconds (2 periods), so I had to fix that.

While at it, another issue was taken care of - namely the tooltip duration was only
using the period duration instead of the whole (dynamic) system activation duration.

That one is also taken care of.

Finally, the new JAR is also put in.