Fix new condition issue and rename it while at it
aka "pay the price for your naivety". Somehow I naively thought
that just making a new class is going to make it all work out magically.

Without registering the new class anywhere, or instantiating a new instance
or anything.

Well, the following crash proved me otherwise:
java.lang.NullPointerException: Condition for ammoWeapons is missing.
	at exoticatechnologies.modifications.conditions.ConditionDict$Companion.getCondFromDict(ConditionDict.kt:70)
	at exoticatechnologies.modifications.conditions.ConditionDict$Companion.getCondFromJSONObject(ConditionDict.kt:50)
	at exoticatechnologies.modifications.conditions.ConditionDict$Companion.getCondsFromJSONArray(ConditionDict.kt:43)
	at exoticatechnologies.modifications.Modification.<init>(Modification.kt:40)
	at exoticatechnologies.modifications.exotics.Exotic.<init>(Exotic.kt:28)
	at exoticatechnologies.modifications.exotics.impl.AmmoCraftingSystem.<init>(AmmoCraftingSystem.kt:30)
	at exoticatechnologies.modifications.exotics.ExoticsHandler.populateExotics(ExoticsHandler.kt:59)
	at exoticatechnologies.modifications.exotics.ExoticsHandler.initialize(ExoticsHandler.kt:36)
	at exoticatechnologies.ETModPlugin.onApplicationLoad(ETModPlugin.java:59)
	at com.fs.starfarer.loading.ResourceLoaderState.init(Unknown Source)

So, lets just register the AmmoWeapons in this ConditionDict.
But also, while creating the PR i got an idea about leaving a comment to rename it to
AmmoUsingWeapons instead, as that much better conveys both it's purpose and what it does.

Also changed it's usage to "ammoUsingWeapons".

EDIT: also added a new JAR while at it.

So I renamed it, updated the exotics.json, and it's ready for another round of testing after a new deployment.