Fix a whoopsie and also fix the image too
Turns out, somehow I didn't add an entry for the AmmoCraftingSystem in
the settings.json. Alright, so I added it.

But while at it, I also removed the background from the old ammocreator (?)
image, and named it properly - ammocraftingsystem.png to be inline with other
images named exactly like their systems are.