Refix the rolledChance again
To make it distribute random numbers in a [0,99] range, we need to add +1
to the 'until' part of the range, because in the random.nextInt(from,until)
the 'from' is inclusive but the 'until' is exclusive. So it does need to
be random.nextInt(0,100) and I reduced it to avoid rolling 101 number instead.