'Fix' the explosion location randomness
So, as it turns out, segments don't really return an absolute location,
but rather just a relative location. However, when we do something like

    Vector2f(
        shipLocation.x + segmentLocation.x,
        shipLocation.y + segmentLocation.y
    )

It worked great for large ships, but worked like crap for the small ship i'm
testing with. With the small ship, the explosion is literally on the edge of
if the ship, or it's corner or something like that.

So, what I think needs to be done is, to get a random (X,Y) in segment's (x/2,y/2)
and go from -that to +that and add that to ship's location. That might work out better.