Attempy to fix segment randomness - try3
Alright, so I see a problem.

[AmmoCraftingSystem] --> generateRandomLocationOnShip()	shipLocation: Vector2f[204.9541, -1535.2051]
[AmmoCraftingSystem] generateRandomLocationOnShip()	segmentLocation: Vector2f[-48.907772, -1629.393]
[AmmoCraftingSystem] generateRandomLocationOnShip()	segment location != ship location
[AmmoCraftingSystem] generateRandomLocationOnShip()	segmentX: -24.453886, segmentY: -814.6965
[AmmoCraftingSystem] generateRandomLocationOnShip()	randX: 98.74124, randY: -816.4
[AmmoCraftingSystem] <-- generateRandomLocationOnShip()	retVal: Vector2f[98.74124, -816.4]

ok, so it seems like i'm still using segmentX/2 thinking that segment.location is relative instead of absolute....
which is obviously wrong.

So, the randomness that *needs* to be done, if any is to be done, is between
shipLocation.x and shipLocation.x + diffX
where diffX = abs(shiplocation.y) - abs(segmentLocation.y)

or abs(segment) - abs(shipLocation), whatever. segments left of the center will produce negative diff, segments
right of the center will produce positive diffs, yet - it doesn't matter where we show the explosion.
we don't need to show it *exactly* on the random segment. we just want them to be within ship bounds and not
always be in the center.

So lets try that idea. But first - lets get rid of segmentX = segmentLocation.x / 2 since that introduces these huge jumps into emptiness.