Attempt to fix segment randomness - try5
Ok, there's no way that some segment locations aren't relative, it just doesn't make sense.
When i test with the "oh pawlease" ship, some segments are 1000+ units away from the ship's location.
That's like more than half of the screen away. can't be true.

So, I either have math that works great for big ships and sucks for this one, or I have math that works good
for the small ship but somehow sucks for the big ship.

So now, I'm doing something different. if the diff > 1000 i "change my mind" and consider the location to be
relative, and recalibrate the diffX and diffY.

Lets see where that gets me so far...