Fix the 'inverted' problem with segment randomness
Well, it turns out, the ship location isn't always at the center, and the whole
"it won't matter if i don't get the sign right" was a big problem when apparently
a segment from the top of the ship was selected and we went that far downwards.

So after the calculation was corrected, I thought the brokenness went away however...
I saw another repro just now, though not as bad as it used to be.

Lets keep this and call it done.