Finally fix the segment problem of spawning explosions/visuals outside of the ship
Well, problem is finally solved for good. Instead of the dumb "switch to absolute positions if diff > 1000"
which didn't make sense to begin with becase:
1) the segment locations are always absolute
2) it's not always going to be 1000 and is most likely ship size dependant

instead of introducing a yet-another-codesmell with comparisons with ship width/height, lets just check if the
generated point is within bounds, and if it's not - generate a new one that is within bounds.

On the downside - this makes the explosions hit other modules less likely (read: impossible)
On the upside - we won't be having explosions outside of ship bounds, period.

The multi-module stuff will be taken care of eventually, in the multimodule themed PR.