Tweak fail damage parameters from 10-100 to 100-400
Changed the 10-100 to 100-400 which seemed very fine during testing.
It does leave some damage after 4-5 activations because the first activations
strip off armor, but in a real-world scenario where the ship is fighting - i guess
we'll just have to see what's there to see and fix later if it needs fixing.