Fix for GuardianShield remaining when owner-ship retreats
Turns out, as reported to me on discord, sometimes when the enemy's ship that had
GuardianShield installed retreats, the drone just remains there with the shield turned on.

So the idea is to do a few things:
1) make the drone explode when the ship dies to avoid having this issue when the ship dies.
2) remove the FXdrone when the ship dies.
3) if the owner-ship is not in play, remove the FXdrone from play if it's present
4) if the owner-ship is in either player's or enemy's retreated ships list - remove the FX drone if present.

Hopefully that takes care of the problem which didn't have a 100% repro rate to begin with.