Fix for buying exoticas which you cannot afford
While the logic was there to check if you can afford it, it turns out that
if you can't afford it - the items are simply desintegrated into nothing,
which can kinda suck if/when a rare exotica shows up in there and you just
poof it into nothingness due to not reading the text and the button letting
you click on it.

Now, if you can't afford the items, they are simply returned back to the market.

Ideally, the button would be grayed out as well, however I haven't figured out
how to do that yet. But this is good enough really.