Framework improvements - part 1
It's obvious that the advanceInCombatAlways() isn't always being called,
since it's being called by ExoticaEveryFramePlugin only while the ship is there.

Meaning, once it leaves - it will no longer be called. Once it dies - it will no longer be called.

Which are good things, but if I'm just here sitting and waiting to be notified of something
it's pretty bad that I will never be notified of anything while relying on it to be called.

So, some refactoring is going on, and Exotic as well as Upgrade received two new methods:
onOwnerShipRemovedFromCombat(..., reason) and onOwnerShipEnteredCombat(...)

'reason' being this new ExoticaShipRemovalReason { NONE, DEATH, RETREAT } enum I forgot to commit.