Resolve the issue of floating GuardianShield after ship retreats in sim
While I haven't seen this problem *many* times, I did see it, and after I could reliably
reproduce it in simulation, it was time to fix it.

It stems from the fact that, before the ship flees (but not always) the drone does not flee,
which doesn't cause that much problem unless the ship decides to order it to bring up the shield.

When it does bring up a shield, we have an empty shield bubble with a huge amount of flux capacity
to shield others, and besides being a big visual issue it could also be a very bad somewhat gamebreaking
thing - because the ship that should get overloaded as the shield tanks up huge amounts of flux will never
get overloaded.

So it's best to get rid of this edge-case completely.

That was done by extending the framework with two more callbacks, and tweaking the GuardianShield to preserve
it's subsystem as well as a "turned off for good" flag, which, when turned on prevents the system from being
turned on.

When the ownerShip gets removed from combat, we remove the shield-hosting drone and turn the system off for good.
When the ownerShip enters combat, we clear the "turned off for good" flag, and everything is back to normal.