Fix issue #10 - upgrades missing description for last level (maxLevel)
Issue link: https://github.com/RkShaRkz/ExoticaTechnologies/issues/10

There was a problem with how the description range was being calculated,
PMD (scardwell) used a range with 'until' which excludes the 'to' range.

So this piece of code

 for (startingLevel in 0 until maxLevel)

was only relevant for [0,10) range, meaning 0 through 9.

Changed the range to not use 'until' but rather '..' operator, which now
correctly includes the 'to' range as well, and covers the whole [0,10] range