Framework update 1 - allow finer-grained control over Modifications on modules
In multi-module environments, the shouldAffectModule() calls controlled too much
it was both in control over whether the exotica/upgrade was installable on modules
**and** whether it's effects would replicate to all other modules of the ship as well.

On top of that, the weird behaviour differences between simulation and real combat is
throwing another wrench in this that went undetected for a long time;

Namely, in simulations, the modifications seemed to only affect modules they were installed on.
In real combat, the modifications seemed to affect and replicate it's effects on all other modules.

So, normally, to cover a edge-case of "being installable on any module but working just on it" had
to be allowed, and that required extending the framework a bit, which was accomplished by introducing
a new shouldShareEffectToOtherModules() method which defaults to 'false'.

While this will break the existing behaviour of all other exoticas/upgrades, it also fixes the problem
of GuardianShield being spawned on all other modules regardless of where it was installed (didn't verify this)
but installing it on the main module certainly replicated the effects to all other modules as well.

Now with this method in place, the doubly-locked spawning introduced to GuardianShield can remain, altho unnecessary
and a pass through all other exoticas/upgrades need to be made to further diversify them as well as update their
descriptions to mention whether they apply to all modules of the ship, or just the module they're installed on.

So far I'm thinking, modifications with penalties should affect the whole ship, and modifications without penalties
should work just where they're installed, to make overpowering your ship harder and more demanding.

For now, this method concerns only where (on which modules) ExoticaTechHM will call the following methods:
- modification.applyToShip
- modification.advanceInCombatUnpaused

while the modification.applyModificationToStats() and modification.applyToFighters() still don't depend on the new method (why?)

Apart from that, some javadoc improvements were made, and I'm terribly disappointed at KDoc's inability to reference
concrete overloaded methods in the @see section, or the general documentation section as well. Can't believe Java has an upper hand here.