Framework update 2 - extend the usage of shouldShareEffectToOtherModules()
In the previous commit, we only allowed the shouldShareEffectToOtherModules() to affect
calls to two methods - applyToShip() and advanceInCombatUnpaused() - while leaving
the applyModificationToStats() and applyToFighters() to remain unaffected by it.

This commit changes that and streamlines the usage of the shouldShareEffectToOtherModules()
to be relevant in calling *all* of the 4 methods from ExoticaTechHM for ships/modules
depending on it's return value, which now defaults to 'false'.

So this means that, with this change, each modification will work only on the module it's
installed on until they are all changed to either work everywhere or remain as-is, relavant to single module.

Apart from that, javadoc got some more love and clarifications which might help in the future.